+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * 【core   】      重载subtractToList方法，提供isLinked选项（pr#3923@Github）
 * 【extra  】      `TemplateConfig`增加`setUseCache`方法（issue#IC3JRY@Gitee）
 * 【extra  】      `AbstractFtp`增加`rename`方法（issue#IC3PMI@Gitee）
+* 【core   】      优化`PropDesc`缓存注解判断，提升性能（pr#1335@Gitee）
 
 ### 🐞Bug修复
 * 【setting】      修复`Setting`autoLoad可能的加载为空的问题（issue#3919@Github）